+++ CHANGELOG.md
@@ -7,3 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2020-04-30
 ### Added
 - CHANGELOG version 1.0.0 release
+
+## [1.1.0] - 2020-06-30
+### Added
+- Elemental Link as an input option
+- changed the MediaLive Encoding segment length from 10 seconds to 4